Handle not found case 1. Get 2. Delete 3. Put 4. Add H2EmployeeDataSetupExtension for handling data setup in junit

+++ src/main/java/jp/co/axa/apidemo/services/EmployeeServiceImpl.java
@@ -4,12 +4,14 @@ import java.util.List;
 import java.util.Optional;
 import jp.co.axa.apidemo.entities.Employee;
 import jp.co.axa.apidemo.exception.EmployeeAlreadyExistsException;
+import jp.co.axa.apidemo.exception.EmployeeNotFoundException;
 import jp.co.axa.apidemo.repositories.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class EmployeeServiceImpl implements EmployeeService {
@@ -30,10 +32,10 @@ public class EmployeeServiceImpl implements EmployeeService {
 
   public Employee getEmployee(Long employeeId) {
     Optional<Employee> optEmp = employeeRepository.findById(employeeId);
-    // FIXME: Handle if employee is not found
-    return optEmp.get();
+    return optEmp.orElseThrow(() -> new EmployeeNotFoundException(employeeId));
   }
 
+  @Transactional
   @Override
   public Employee saveEmployee(Employee employee) {
     String employeeNumber = employee.getNumber();
@@ -43,11 +45,25 @@ public class EmployeeServiceImpl implements EmployeeService {
     return employeeRepository.save(employee);
   }
 
+  @Transactional
+  @Override
   public void deleteEmployee(Long employeeId) {
-    employeeRepository.deleteById(employeeId);
+    if (employeeRepository.existsById(employeeId)) {
+      employeeRepository.deleteById(employeeId);
+    } else {
+      throw new EmployeeNotFoundException(employeeId);
+    }
   }
 
-  public void updateEmployee(Employee employee) {
-    employeeRepository.save(employee);
+  @Transactional
+  @Override
+  public Employee updateEmployee(Long employeeId, Employee employee) {
+    if (employeeRepository.existsById(employeeId)) {
+      // in case user provide a different id.
+      employee.setId(employeeId);
+      return employeeRepository.save(employee);
+    } else {
+      throw new EmployeeNotFoundException(employeeId);
+    }
   }
 }
